Parameterize base test file

+++ medusa_test.sh
@@ -1,28 +1,49 @@
 #!/bin/bash
 
-WSS=2048
+if [ "$#" -ne 3 ]; then
+    echo "Usage: $0 <co/no> <WSS(KB)> <corun-type(read/write)>"
+    exit 1
+fi
+
+corun=$1
+WSS=$2
+corun_type=$3
+
+type="Solo"
+
 apart=part1 # part 1 == bin 0
 vpart=part1 # part9 == bins 1-3
 
-killall bandwidth
-killall latency
 # start co-runners on cores 1,2,3
 # with WSS > LLC_SIZE to force requests to go to memory
 
-for i in 1 2 3; do
-    ./bandwidth -c $i -m $WSS -t 0 -a write &
-    bw_pid=$!
-    echo $bw_pid
-    echo $bw_pid >> /sys/fs/cgroup/palloc/$apart/tasks
-done
+> /sys/fs/cgroup/palloc/$vpart/cgroup.procs
 
-echo "WS(KB)    latency(ns)"
-./latency -c 0 -m $WSS -i 100 > out.txt 2> /dev/null &
+if [ "$corun" = "co" ]; then
+    echo "Beginning corunners...."
+    type="Corun"
+    for i in 1 2 3; do
+        ./bandwidth -c $i -m 2048 -t 0 -a $corun_type > /dev/null 2>&1 &
+        bw_pid=$!
+        echo $bw_pid
+        echo $bw_pid >> /sys/fs/cgroup/palloc/$apart/cgroup.procs
+    done
+fi
+
+echo -e "\nVictim running...."
+
+chrt -f 99 ./latency -c 0 -m $WSS -i 100 > out.txt 2> /dev/null &
 latency_pid=$!
-chrt -f -p 99 $latency_pid
-echo $latency_pid > /sys/fs/cgroup/palloc/$vpart/tasks
+echo $latency_pid >> /sys/fs/cgroup/palloc/$vpart/cgroup.procs
 echo $latency_pid
 wait $latency_pid
 
+if [ "$corun" = "co" ]; then
+    echo "Killing co-runners...."
+    killall bandwidth
+    wait &> /dev/null
+fi
+
 VAL=`grep average out.txt | awk '{ print $2 }'`
-echo -e $ws '\t' $VAL
+echo -e "\nType\tWSS(KB)\tVictim latency(ns)"
+echo -e "$type\t$WSS\t$VAL"
